fix(quota): widen token-quota columns INT4→BIGINT (fixes self-service "更新失敗" on ≥2.1e9)
Setting a self-service default quota of 10e9 tokens 500'd with "更新失敗": the value
overflowed the Postgres INTEGER (INT4, max ~2.147e9) column. Three token-quota columns
were INT4 — allocations.quota_tokens_per_month, model_catalog.self_service_default_quota,
pool_config.total_tokens_per_month — so even if the model config saved, the later
self-service *claim* (which copies the quota to an allocation) would overflow too.

Fix: widen all three to BigInteger (migration 0023, dialect-guarded — SQLite is already
64-bit + can't ALTER TYPE). Added a sane API cap (default_quota ≤ 1e15) so pathological
values 422 cleanly instead of erroring at the DB. SQLite stores 64-bit regardless, so the
regression guard asserts the column *types* are BigInteger. 823 tests, ruff/mypy clean.

+++ src/ai_api/models/model_catalog.py
@@ -9,7 +9,7 @@ import enum
 from datetime import datetime
 from typing import Any
 
-from sqlalchemy import JSON, Boolean, DateTime, Enum, Index, Integer, String, Text
+from sqlalchemy import JSON, BigInteger, Boolean, DateTime, Enum, Index, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from ai_api.db import Base
@@ -53,7 +53,7 @@ class ModelCatalog(Base):
 
     # Phase 6: self-service allocation
     self_service_enabled: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    self_service_default_quota: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    self_service_default_quota: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
 
     # Phase 23: LiteLLM registry provenance — {base_model_key, imported_version,
     # field_sources: {field: litellm|borrowed|manual}, snapshot: {field: value}}.
